Input style attributes are added and cleaned up code
