+++ src/components/index.js
@@ -1,2 +1,4 @@
-export * from './renderInputDetails';
 export * from './inputWithLabel';
+export * from './ButtonDetails';
+export * from './InputDetails';
+export * from './Navbar'
