+++ src/components/inputWithLabel.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-export const InputWithLabel = ({label, inputName, inputValue, onChangeHandler, inputType = "text"}) => {
+export const InputWithLabel = ({label, inputName, inputValue, onChangeHandler, inputType = "text", key}) => {
   return (
     <>
         <label className="justify-self-end self-center">{label}</label>
@@ -10,6 +10,7 @@ export const InputWithLabel = ({label, inputName, inputValue, onChangeHandler, i
             className="p-2 border-2 border-solid rounded-md w-48 sm:w-80"
             value={inputValue}
             onChange={onChangeHandler}
+            required
         />
     </>
   )
