+++ src/constants/index.js
@@ -1,1 +1,2 @@
 export * from './styleOptions';
+export * from './initialValues';
