+++ src/constants/initialValues.js
@@ -1,0 +1,16 @@
+export const BUTTON_INITIAL_VALUES = {
+    name: "",
+    tag: "",
+    type: "",
+    value: "",
+    styles: []
+}
+
+export const INPUT_INITIAL_VALUES = {
+    name: "",
+    tag: "",
+    type: "",
+    value: "",
+    placeholder: "",
+    styles: []
+}
