+++ src/main.css
@@ -6,3 +6,6 @@
   font-family: sans-serif;
 }
 
+input[type="checkbox"]:checked {
+  accent-color: #9333ea;
+}
