+++ src/pages/Elements.jsx
@@ -2,6 +2,9 @@ import React from 'react'
 
 export const Elements = () => {
   return (
-    <div className="flex-grow text-center font-medium py-6">Will Render the created UI elements here</div>
+    <div className="flex-grow text-center font-medium py-6">
+      <p>Will Render the created UI elements here</p>
+      <p>The user can edit/ delete from this page</p>
+    </div>
   )
 }
